There are mainly four points as the following:
1.There are two versions of "Stochastic Transparency": "Basic" and "Depth-Based". We only provide the "Depth-Based" version.

2.We rename the related names to be consistent with The paper.

3.We calculate Vis(z) in AccumulationPass.
The original code provided by the NVIDIA test against the StochasticDepth in AccumulationPass instead.

4."Turning on MSAA in StochasticDepthPass is to random sample and the stochastic transparency intrinsically doesn't demand other passes to turn on the MSAA."
The original code provided by the NVIDIA turns on the MSAA in all passes and thus the frame rate of the stochastic transparency is unexpectedly lower than the depth peeling.
I turn off all the unnecessary MSAA and the frame rate increases from 670 to 1170 while the frame rate of deep peeling is 1070.